abstracted transfer into internal method, removed all warnings
